// src/constants/enums.ts
export enum TASK_STATUS {
    DONE = "DONE",
    IN_PROGRESS = "IN_PROGRESS",
    TO_DO = "TO_DO"
}

export enum ErrorType {
    VALIDATION_ERROR = 'VALIDATION_ERROR',
    BUSINESS_ERROR = 'BUSINESS_ERROR',
    AUTHENTICATION_ERROR = 'AUTHENTICATION_ERROR',
    AUTHORIZATION_ERROR = 'AUTHORIZATION_ERROR',
    RESOURCE_NOT_FOUND = 'RESOURCE_NOT_FOUND',
    RATE_LIMIT_ERROR = 'RATE_LIMIT_ERROR',
    DEPENDENCY_ERROR = 'DEPENDENCY_ERROR',
    SYSTEM_ERROR = 'SYSTEM_ERROR',
}